+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This page reports the main releases only and the main changes therein.
 * more safety checks prevent crashes for abnormal conditions
 * support reference_index = -1 as last index; indices renumbered starting from 1; 0 means media index.
 * improved reference frame GUI widget
+* improved project undo description
 * code cleanup
 
 ---